shill: network: Make integer conversion explicit
Also use int instead of uint_32 for interface indices for consistency
with other part of code base.

BUG=b:293997937
TEST=unit

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5119140
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Tested-by: Taoyu Li <taoyl@chromium.org>

+++ shill/network/network_applier.cc
@@ -302,7 +302,7 @@ void NetworkApplier::ApplyRoutingPolicy(
 }
 
 void NetworkApplier::ApplyMTU(int interface_index, int mtu) {
-  rtnl_handler_->SetInterfaceMTU(interface_index, mtu);
+  rtnl_handler_->SetInterfaceMTU(interface_index, static_cast<uint32_t>(mtu));
 }
 
 void NetworkApplier::ApplyRoute(
